Renamed src folder to sim_game (python package)
