+++ tests/path_fix.py
@@ -4,5 +4,5 @@
 import sys
 import os
 sys.path.insert(0, "./")
-if "/src" not in os.getcwd() and os.path.exists("./src/__init__.py"):
-    sys.path.insert(0, "./src")
+if "/sim_game" not in os.getcwd() and os.path.exists("./sim_game/__init__.py"):
+    sys.path.insert(0, "./sim_game")
