+++ tests/test_import.py
@@ -3,7 +3,7 @@
 import path_fix
 
 def test_import():
-    from src import game
-    from src import ci_dummy
-    from src import main
-    from src.ci_dummy import double_a_number
+    from sim_game import game
+    from sim_game import ci_dummy
+    from sim_game import main
+    from sim_game.ci_dummy import double_a_number
